+++ test/test-extra-urlite.js
@@ -1,20 +1,18 @@
 /* globals describe it */
 var expect = require('expect.js')
-var urlite = require('../querystring/urlite')
+var urlite = require('../extra')
 
 describe('querystring urlite', function () {
-  it('should parse the querystring', function () {
+  it('should parse extras', function () {
     var url = 'proto://user:password@domain.com:3000/some/pathname?a=b#fragment'
     var parsed = {
-      auth: 'user:password',
+      auth: { user: 'user', password: 'password' },
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
       pathname: '/some/pathname',
       path: '/some/pathname?a=b',
-      search: '?a=b',
-      query: {
+      search: {
         a: 'b'
       },
       hash: '#fragment',
@@ -27,33 +25,47 @@ describe('querystring urlite', function () {
   it('should handle no query', function () {
     var url = 'proto://user:password@domain.com:3000/some/pathname#fragment'
     var parsed = {
-      auth: 'user:password',
+      auth: { user: 'user', password: 'password' },
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
       pathname: '/some/pathname',
       path: '/some/pathname',
       search: undefined,
-      query: undefined,
       hash: '#fragment',
       href: url
     }
     expect(urlite.parse(url)).to.eql(parsed)
     expect(urlite.format(parsed)).to.eql(url)
   })
-  it('should handle empty querystring params', function () {
-    var url = 'proto://user:password@domain.com:3000/some/pathname?blah&blah2=&boop'
+
+  it('should handle no auth', function () {
+    var url = 'proto://domain.com:3000/some/pathname#fragment'
     var parsed = {
-      auth: 'user:password',
+      auth: undefined,
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
+      pathname: '/some/pathname',
+      path: '/some/pathname',
+      search: undefined,
+      hash: '#fragment',
+      href: url
+    }
+    expect(urlite.parse(url)).to.eql(parsed)
+    expect(urlite.format(parsed)).to.eql(url)
+  })
+
+  it('should handle empty querystring params', function () {
+    var url = 'proto://user:password@domain.com:3000/some/pathname?blah&blah2=&boop'
+    var parsed = {
+      auth: { user: 'user', password: 'password' },
+      protocol: 'proto:',
+      port: '3000',
+      hostname: 'domain.com',
       pathname: '/some/pathname',
       path: '/some/pathname?blah&blah2=&boop',
-      search: '?blah&blah2=&boop',
-      query: {
+      search: {
         blah: '',
         blah2: '',
         boop: ''
@@ -66,15 +78,13 @@ describe('querystring urlite', function () {
   it('should handle empty querystring values', function () {
     var url = 'proto://user:password@domain.com:3000/some/pathname?=&'
     var parsed = {
-      auth: 'user:password',
+      auth: { user: 'user', password: 'password' },
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
       pathname: '/some/pathname',
       path: '/some/pathname?=&',
-      search: '?=&',
-      query: {
+      search: {
         '': ''
       },
       hash: undefined,
