+++ test/test-format.js
@@ -10,7 +10,6 @@ describe('format', function () {
       protocol: undefined,
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: undefined,
       search: undefined,
       hash: undefined,
@@ -25,7 +24,6 @@ describe('format', function () {
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
       pathname: '/some/pathname',
       search: '?query=string',
       hash: '#fragment',
@@ -40,7 +38,6 @@ describe('format', function () {
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
       pathname: '/some/pathname',
       search: '?query=string',
       hash: '#fragment',
@@ -55,7 +52,6 @@ describe('format', function () {
       protocol: undefined,
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: '/some/pathname',
       search: '?query=string',
       hash: '#fragment',
@@ -70,7 +66,6 @@ describe('format', function () {
       protocol: undefined,
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: '/some/pathname',
       search: undefined,
       hash: '#fragment',
@@ -85,7 +80,6 @@ describe('format', function () {
       protocol: undefined,
       port: undefined,
       hostname: 'example.com',
-      host: 'example.com',
       pathname: '/hello',
       search: undefined,
       hash: undefined,
