+++ test/test-parse.js
@@ -9,11 +9,9 @@ describe('parse', function () {
       protocol: undefined,
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: undefined,
       path: undefined,
       search: undefined,
-      query: undefined,
       hash: undefined,
       href: ''
     })
@@ -26,11 +24,9 @@ describe('parse', function () {
       protocol: 'proto:',
       port: '3000',
       hostname: 'domain.com',
-      host: 'domain.com:3000',
       pathname: '/some/pathname',
       path: '/some/pathname?query=string',
       search: '?query=string',
-      query: 'query=string',
       hash: '#fragment',
       href: url
     })
@@ -41,14 +37,12 @@ describe('parse', function () {
     expect(parse(url)).to.eql({
       auth: 'user:password',
       hash: '#fragment',
-      host: 'domain.com:3000',
       hostname: 'domain.com',
       href: url,
       path: '/some/pathname?query=string',
       pathname: '/some/pathname',
       port: '3000',
       protocol: 'proto:',
-      query: 'query=string',
       search: '?query=string'
     })
   })
@@ -60,11 +54,9 @@ describe('parse', function () {
       protocol: undefined,
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: '/some/pathname',
       path: '/some/pathname?query=string',
       search: '?query=string',
-      query: 'query=string',
       hash: '#fragment',
       href: url
     })
@@ -77,11 +69,9 @@ describe('parse', function () {
       protocol: undefined,
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: '/some/pathname',
       path: '/some/pathname',
       search: undefined,
-      query: undefined,
       hash: '#fragment',
       href: url
     })
@@ -98,8 +88,6 @@ describe('parse', function () {
       hostname: 'dev---www-sitting--duck-com.poxy.com',
       auth: undefined,
       protocol: 'http:',
-      host: 'dev---www-sitting--duck-com.poxy.com:666',
-      query: '_q_portal_protocol=http',
       path: '?_q_portal_protocol=http'
     })
   })
@@ -111,11 +99,9 @@ describe('parse', function () {
       protocol: 'javascript:',
       port: undefined,
       hostname: undefined,
-      host: undefined,
       pathname: 'alert("node is awesome");',
       path: 'alert("node is awesome");',
       search: undefined,
-      query: undefined,
       hash: undefined,
       href: url
     })
@@ -132,8 +118,6 @@ describe('parse', function () {
       hostname: 'localhost',
       auth: undefined,
       protocol: 'http:',
-      host: 'localhost:4400',
-      query: undefined,
       path: '/@qubit/layer@2.20.47/lib/layer_editor.js'
     })
   })
@@ -144,12 +128,10 @@ describe('parse', function () {
       result: {
         protocol: 'http:',
         auth: undefined,
-        host: 'nodejs.org',
         port: undefined,
         hostname: 'nodejs.org',
         hash: '#url_url_format_urlobj',
         search: undefined,
-        query: undefined,
         pathname: '/docs/latest/api/url.html',
         path: '/docs/latest/api/url.html',
         href: 'http://nodejs.org/docs/latest/api/url.html#url_url_format_urlobj'
@@ -159,12 +141,10 @@ describe('parse', function () {
       result: {
         protocol: 'http:',
         auth: undefined,
-        host: 'blog.nodejs.org',
         port: undefined,
         hostname: 'blog.nodejs.org',
         hash: undefined,
         search: undefined,
-        query: undefined,
         pathname: '/',
         path: '/',
         href: 'http://blog.nodejs.org/'
@@ -174,12 +154,10 @@ describe('parse', function () {
       result: {
         protocol: 'https:',
         auth: undefined,
-        host: 'encrypted.google.com',
         port: undefined,
         hostname: 'encrypted.google.com',
         hash: undefined,
         search: '?q=url&q=site:npmjs.org&hl=en',
-        query: 'q=url&q=site:npmjs.org&hl=en',
         pathname: '/search',
         path: '/search?q=url&q=site:npmjs.org&hl=en',
         href: 'https://encrypted.google.com/search?q=url&q=site:npmjs.org&hl=en'
@@ -189,12 +167,10 @@ describe('parse', function () {
       result: {
         protocol: undefined,
         auth: undefined,
-        host: undefined,
         port: undefined,
         hostname: undefined,
         hash: '#whoo',
         search: '?oh=yes',
-        query: 'oh=yes',
         pathname: 'some.ran/dom/url.thing',
         path: 'some.ran/dom/url.thing?oh=yes',
         href: 'some.ran/dom/url.thing?oh=yes#whoo'
@@ -204,12 +180,10 @@ describe('parse', function () {
       result: {
         protocol: 'https:',
         auth: 'user:pass',
-        host: 'example.com',
         port: undefined,
         hostname: 'example.com',
         hash: undefined,
         search: undefined,
-        query: undefined,
         pathname: '/',
         path: '/',
         href: 'https://user:pass@example.com/'
@@ -219,12 +193,10 @@ describe('parse', function () {
       result: {
         protocol: undefined,
         auth: undefined,
-        host: undefined,
         port: undefined,
         hostname: undefined,
         hash: undefined,
         search: undefined,
-        query: undefined,
         pathname: '/wiki/Help:IPA',
         path: '/wiki/Help:IPA',
         href: '/wiki/Help:IPA'
@@ -234,12 +206,10 @@ describe('parse', function () {
       result: {
         protocol: 'http:',
         auth: ':pass',
-        host: 'example.org:123',
         port: '123',
         hostname: 'example.org',
         hash: '#fragment',
         search: '?query=string',
-        query: 'query=string',
         pathname: '/some/directory/file.html',
         path: '/some/directory/file.html?query=string',
         href: 'http://:pass@example.org:123/some/directory/file.html?query=string#fragment'
